requirements.txt
Flask==1.1.2
pymongo==3.11.2
gunicorn
python-dotenv